+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All of the packages in the `apollo-server` repo are released with the same versi
 
 ### v2.0.4
 
+- Core: Allow context to be passed to all GraphQLExtension methods. [PR #1547](https://github.com/apollographql/apollo-server/pull/1547)
 - apollo-server: Release due to failed build and install
 
 ### v2.0.3